chore: refactor and speed up resetting scenarios
