+++ e2e/home.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test';
-import { TEST_IDS } from '../src/client/e2eTestIDs';
+import { TEST_IDS } from '../src/client/testIDs';
 
 test.describe('Home page', () => {
   test('should list cards with use-cases', async ({ page }) => {
