+++ e2e/scraping/protected.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test';
-import { TEST_IDS } from '../../src/client/e2eTestIDs';
+import { TEST_IDS } from '../../src/client/testIDs';
 
 test.describe('Scraping flights', () => {
   test('is not possible with Bot detection on', async ({ page }) => {
