+++ src/client/components/personalization/searchComponents.tsx
@@ -2,7 +2,7 @@ import { FunctionComponent } from 'react';
 import Image from 'next/image';
 import SearchIcon from '../../img/search.svg';
 import styles from './searchComponents.module.scss';
-import { TEST_IDS } from '../../e2eTestIDs';
+import { TEST_IDS } from '../../testIDs';
 
 type SearchProps = {
   search: string;
